# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vaoucher Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/reqMasterData": {
            "post": {
                "summary": "Submit master data",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReqMasterData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Master Data Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/detailsData": {
            "post": {
                "summary": "Submit details data",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetailsData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Details Data Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/fileDetails": {
            "post": {
                "summary": "Submit file details",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FileDetails"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File Details Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/reqHistory": {
            "post": {
                "summary": "Submit requisition history",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReqHistory"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Requisition History Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/reqStatus": {
            "post": {
                "summary": "Submit requisition status",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReqStatus"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Requisition Status Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/status": {
            "post": {
                "summary": "Submit status",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Requisition Status Submitted!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/allData": {
            "get": {
                "summary": "Retrieve all requisition data",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RequisitionData"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/comments": {
            "post": {
                "summary": "Submit comments",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "req_id": {
                                        "type": "string"
                                    },
                                    "comments": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "user_id",
                                    "req_id",
                                    "comments"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comments submitted"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ReqMasterData": {
                "type": "object",
                "properties": {
                    "requisition_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "customer_id": {
                        "type": "string"
                    },
                    "project_id": {
                        "type": "string"
                    },
                    "supplier_id": {
                        "type": "string"
                    },
                    "total_amount": {
                        "type": "number"
                    }
                }
            },
            "DetailsData": {
                "type": "object",
                "properties": {
                    "requisition_details_id": {
                        "type": "string"
                    },
                    "requisition_master_id": {
                        "type": "string"
                    },
                    "item_category_id": {
                        "type": "string"
                    },
                    "item_name": {
                        "type": "string"
                    },
                    "uom": {
                        "type": "string"
                    },
                    "qty": {
                        "type": "number"
                    },
                    "unit_price": {
                        "type": "number"
                    }
                }
            },
            "FileDetails": {
                "type": "object",
                "properties": {
                    "requisition_file_id": {
                        "type": "string"
                    },
                    "requisition_master_id": {
                        "type": "string"
                    },
                    "file_url": {
                        "type": "string"
                    }
                }
            },
            "ReqHistory": {
                "type": "object",
                "properties": {
                    "history_id": {
                        "type": "string"
                    },
                    "requisition_master_id": {
                        "type": "string"
                    },
                    "action_done_by": {
                        "type": "string"
                    },
                    "action_status_id": {
                        "type": "string"
                    }
                }
            },
            "ReqStatus": {
                "type": "object",
                "properties": {
                    "requisition_status_id": {
                        "type": "string"
                    },
                    "requisition_master_id": {
                        "type": "string"
                    },
                    "status_id": {
                        "type": "string"
                    },
                    "assigned_to": {
                        "type": "string"
                    }
                }
            },
            "Status": {
                "type": "object",
                "properties": {
                    "status_id": {
                        "type": "string"
                    }
                }
            },
            "RequisitionData": {
                "type": "object",
                "properties": {
                    "requisition_id": {
                        "type": "integer"
                    },
                    "requisition_details_data": {
                        "$ref": "#/components/schemas/RequisitionDetailsData"
                    },
                    "requisition_file_details": {
                        "$ref": "#/components/schemas/RequisitionFileDetails"
                    },
                    "requisition_history": {
                        "$ref": "#/components/schemas/RequisitionHistory"
                    },
                    "requisition_status": {
                        "$ref": "#/components/schemas/RequisitionStatus"
                    },
                    "comments": {
                        "$ref": "#/components/schemas/Comments"
                    }
                }
            },
            "RequisitionDetailsData": {
                "type": "object"
            },
            "RequisitionFileDetails": {
                "type": "object"
            },
            "RequisitionHistory": {
                "type": "object"
            },
            "RequisitionStatus": {
                "type": "object"
            },
            "Comments": {
                "type": "object"
            },
...
        }
    }
}